Fix styling

+++ migrations/0002_example.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS "example" (
     "id" VARCHAR(20) PRIMARY KEY,
     "name" VARCHAR(100) NOT NULL,
-    "age" INTEGER NOT NULL,
+    "age" BIGINT NOT NULL,
     "country" VARCHAR(2) NOT NULL,
     "role" VARCHAR(50) NOT NULL,
     "settings" JSONB NOT NULL,
